REFACTOR encryption & compression enums/dataclasses into their own files
