+++ pdfconduit/pdfconduit.py
@@ -1,10 +1,6 @@
 import os
-from dataclasses import dataclass
 from datetime import datetime
-from enum import Enum
-from typing import Optional, Any, Tuple, List, Dict
-
-from pdfconduit import Merge
+from typing import Optional, Any, Tuple, Dict
 
 try:
     from typing import Self, Annotated
@@ -12,93 +8,13 @@ except ImportError:
     from typing_extensions import Self, Annotated
 
 from pypdf import PdfWriter, PdfReader
-from pypdf.constants import UserAccessPermissions
 
+from pdfconduit import Merge
+from pdfconduit.compression import Compression, ImageQualityRange
+from pdfconduit.encryption import Encryption
 from pdfconduit.transform import Rotate, Upscale
 from pdfconduit.convert import Flatten
 from pdfconduit.utils import Info, pypdf_reader, add_suffix
-
-
-class Algorithms(Enum):
-    RC4_40: str = "RC4-40"
-    RC4_128: str = "RC4-128"
-    AES_128: str = "AES-128"
-    AES_256: str = "AES-256"
-    AES_256_r5: str = "AES-256-R5"
-
-    @property
-    def bit_length(self):
-        if "40" in self.value:
-            return 40
-        elif "128" in self.value:
-            return 128
-        elif "256" in self.value:
-            return 256
-
-    @property
-    def is_40bit(self) -> bool:
-        return self.bit_length == 40
-
-    @property
-    def is_128bit(self) -> bool:
-        return self.bit_length == 128
-
-    @property
-    def is_256bit(self) -> bool:
-        return self.bit_length == 256
-
-
-@dataclass
-class Encryption:
-    user_pw: str
-    owner_pw: Optional[str]
-    allow_printing: bool = True
-    allow_commenting: bool = False
-    algo: Algorithms = Algorithms.AES_256_r5
-    permissions: UserAccessPermissions = UserAccessPermissions
-
-    def __post_init__(self):
-        if self.allow_printing and self.allow_commenting:
-            self.permissions = (
-                UserAccessPermissions.PRINT | UserAccessPermissions.MODIFY
-            )
-        elif self.allow_printing:
-            self.permissions = UserAccessPermissions.PRINT
-        elif self.allow_commenting:
-            self.permissions = UserAccessPermissions.MODIFY
-        else:
-            self.permissions = UserAccessPermissions.R2
-
-
-class Compression(Enum):
-    DEFAULT: int = -1
-    NONE: int = 0
-    BEST_SPEED: int = 1
-    BEST_COMPRESSION: int = 9
-    LEVEL_0: int = 0
-    LEVEL_1: int = 1
-    LEVEL_2: int = 2
-    LEVEL_3: int = 3
-    LEVEL_4: int = 4
-    LEVEL_5: int = 5
-    LEVEL_6: int = 6
-    LEVEL_7: int = 7
-    LEVEL_8: int = 8
-    LEVEL_9: int = 9
-
-    @classmethod
-    def from_level(cls, level):
-        return cls(level)
-
-    @classmethod
-    def all(cls) -> List[Self]:
-        return list(map(lambda c: c, cls))
-
-
-@dataclass
-class ImageQualityRange:
-    min: int = 1
-    max: int = 99
 
 
 class Conduit:
